requirement.txt
attrs==20.3.0
cached-property==1.5.2
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
dlib==19.21.1
fvcore==0.1.3.post20210204
h5py==3.1.0
imageio==2.9.0
imageio-ffmpeg==0.4.3
imgaug==0.4.0
importlib-metadata==3.10.0
iniconfig==1.1.1
iopath==0.1.3
kiwisolver==1.3.1
matplotlib==3.3.4
natsort==7.1.1
networkx==2.5
numpy==1.20.0
opencv-python==4.5.1.48
packaging==20.9
pandas==1.2.1
Pillow==8.1.0
pluggy==0.13.1
portalocker==2.2.1
progress==1.5
protobuf==3.14.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
scikit-image==0.18.1
scipy==1.6.2
Shapely==1.7.1
six==1.15.0
tabulate==0.8.7
tensorboardX==2.1
termcolor==1.1.0
tifffile==2021.2.1
toml==0.10.2
torchaudio==0.7.2
tqdm==4.56.0
typing-extensions==3.7.4.3
yacs==0.1.8
zipp==3.4.1
